pyproject.toml
[tool.poetry]
name = "pythonproject10"
version = "0.1.0"
description = ""
authors = ["Dmitry_X <dmitrydsm@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pandas-stubs = "^2.2.2.240909"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.8.0"
mypy = "^1.11.2"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest-cov = "^5.0.0"
types-requests = "^2.32.0.20240914"
pytest = "^8.3.3"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | dist
  )/
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
